backend/requirements.txt
django
pytest
psycopg2-binary
redis
numpy
pandas
structlog
scikit-learn
opentelemetry-sdk
opentelemetry-api
dash
djangorestframework-simplejwt
social-auth-app-django